Restructured semantic module and added similarity calculation.
